Added "add institution" feature
Now you can add an institution and it will be treated like the rest of the markers. One issue is that newly added institutions display formatted address using results[0].formatted_address, instead of the name of the institution like the original locations. Need to figure out how to display just name.